@@ -11,13 +11,23 @@ export class MovingEntity extends Entity {
     maxEnergy: number;
     energyGainRate: number;
     energyLossRate: number;
-    traits = [];
     modifiers = [];
     isSleeping: boolean;
+    duplicate: number;
     goals = [];
 
-    constructor(id: string, position: Position, settingsIndex: number, baseSpriteIndex: number) {
-        super('person', id, position, settingsIndex, baseSpriteIndex);
+    constructor(name: string, id: string, position: Position, settingsIndex: number, baseSpriteIndex: number) {
+        super(name, 'moving-entity', id, position, settingsIndex, baseSpriteIndex);
 
     }
+
+    getInfo() {
+        const info = super.getInfo();
+        info.push(`energy: ${this.energy.toFixed(0)}/${this.maxEnergy}`);
+        info.push(`duplicate: ${this.duplicate.toFixed(1)}`);
+        for (const trait of this.traits) {
+            info.push(`trait: ${trait.name} (${trait.amount.toFixed(3)})`);
+        }
+        return info;
+      }
 }
