@@ -9,9 +9,11 @@ var io = socketIO(server);
 var fs = require("fs");
 //simulation classes
 var Map = require('./server/simulation/classes/map-gen');
-var Goap = require('./server/simulation/classes/goap');
+var Goap = require('./server/simulation/classes/goap/goap');
 var Camera = require('./server/simulation/classes/camera');
 var User = require('./server/simulation/classes/user');
+var GoapPlanner = require('./server/simulation/classes/goap/goap-planner');
+var AStar = require('./server/simulation/classes/a-star/a-star');
 
 //global variables
 var saveDir = `./server/simulation/save`;
@@ -21,8 +23,13 @@ var regenerate = true;
 var players = {};
 var users = {};
 var map;
-
+var goapPlanner = new GoapPlanner();
+var aStar = new AStar();
+var oldTime = Date.now();
+var newTime = oldTime + 100 / 6;
 //=========================================
+//Settings
+var maxEntityPerFrame = 100;
 
 //server setup
 app.set('port', 5000);
@@ -63,12 +70,80 @@ if (regenerate) {
 }
 
 //GOAP setup
-var goap = new Goap();
+goapPlanner.goap = new Goap();
 console.log('laoding goap actions');
-goap.loadActions(goapActionsPath).then(() => {
+goapPlanner.goap.loadActions(goapActionsPath).then(() => {
   // console.log(JSON.stringify(goap.actions));
   console.log(`goap action loading complete`);
 });
+
+function createData(map) {
+  map.data = [];
+  map.interval = 60;
+  map.currentLatestTime = 60;
+  map.currentIndex = 0;
+
+  for (const entitySetting of map.entitySettings) {
+    if (!map.data.find(x => x.name === entitySetting.name)) {
+      map.data.push({
+        series: [{ name: 0, value: 0 }],
+        name: entitySetting.name
+      });
+
+      for (const trait of entitySetting.traits) {
+        map.data.push({
+          series: [{ name: 0, value: 0 }],
+          name: `${entitySetting.name} Average ${trait.name}`
+        });
+      }
+    }
+  }
+}
+
+function updateData(map) {
+  let totals = {};
+  for (const entitySetting of map.entitySettings) {
+    if (!totals[entitySetting.name]) {
+      totals[entitySetting.name] = {
+        total: 0,
+        averages: {}
+      };
+
+      for (const trait of entitySetting.traits) {
+        totals[entitySetting.name].averages[trait.name] = 0;
+      }
+    }
+  }
+
+  for (const entity of map.entities) {
+    if (!entity.destroy) {
+      totals[entity.name].total++;
+      for (const trait of entity.traits) {
+        totals[entity.name].averages[trait.name] += trait.amount;
+      }
+    }
+  }
+
+  // console.log(`average age: ${averageAge}`);
+
+  if (map.time > map.currentLatestTime) {
+    map.currentIndex++;
+    map.currentLatestTime += map.interval;
+    let name = (map.currentLatestTime / 60);
+    for (const data of map.data) {
+      data.series.push({ name: name, value: 0 });
+    }
+  }
+
+  for (const entitySetting of map.entitySettings) {
+    map.data.find(x => x.name === entitySetting.name).series[map.currentIndex].value = totals[entitySetting.name].total;
+
+    for (const trait of entitySetting.traits) {
+      map.data.find(x => x.name === `${entitySetting.name} Average ${trait.name}`).series[map.currentIndex].value
+        = totals[entitySetting.name].averages[trait.name] / totals[entitySetting.name].total;
+    }
+  }
+}
 
 // Add the WebSocket handlers
 io.on('connection', function (socket) {
@@ -82,15 +157,15 @@ io.on('connection', function (socket) {
       //create player
       console.log(`new user ${username} logged in`);
       players[socket.id] = username;
-      const x = map.width / 2;
-      const y = map.height / 2;
+      const x = 0;
+      const y = 0;
       console.log(`x: ${x}, y: ${y}`);
-      users[username] = new User(socket.id, username, socket.id, new Camera({x:x, y:y}, 5));
+      users[username] = new User(socket.id, username, socket.id, new Camera({ x: x, y: y }, 5));
     }
 
     io.to(socket.id).emit('user', users[username]);
     io.to(socket.id).emit('map', map);
-    io.to(socket.id).emit('goapActions', goap.actions);
+    io.to(socket.id).emit('goapActions', goapPlanner.goap.actions);
 
   });
   socket.on('camera', function (camera) {
@@ -111,7 +186,20 @@ io.on('connection', function (socket) {
         }
       });
     } else {
-      io.to(socket.id).emit('logout');
+      // io.to(socket.id).emit('logout');
+      // console.log('unrecognised socket id, sent logout to ' + socket.id);
+      let username = 'default';
+      // delete players[users[username].socketId]
+      players[socket.id] = username;
+      console.log(`${username} logged in`);
+      if (!users[username]) {
+        users[username] = new User(socket.id, username, socket.id, new Camera({ x: 0, y: 0 }, 5));
+      }
+      users[username].socketId = socket.id;
+
+      io.to(socket.id).emit('user', users[username]);
+      io.to(socket.id).emit('map', map);
+      io.to(socket.id).emit('goapActions', goapPlanner.goap.actions);
     }
 
     //console.log(`x: ${player.x}, y: ${player.y}, zoom: ${player.z}`)
@@ -124,26 +212,43 @@ io.on('connection', function (socket) {
       let person = map.people.find(x => x.id == user.username);
       person.goals.push(goal);
     } else {
-      io.to(socket.id).emit('logout');
+      // console.log('sent logout to ' + user.username);
+      // io.to(socket.id).emit('logout');
     }
 
     //console.log(`x: ${player.x}, y: ${player.y}, zoom: ${player.z}`)
   });
 });
 
-setInterval(function () {
-  map.run(goap, 1 / 60);
+function run() {
+  newTime = Date.now();
+  let deltaTime = newTime - oldTime;
+  oldTime = newTime;
+  deltaTime /= 1000;
+  goapPlanner.run();
+  map.run(goapPlanner, deltaTime, aStar, maxEntityPerFrame);
+  // console.log(`sending entity: ${JSON.stringify(map.entities[1])}`);
   io.sockets.emit('entities', map.entities);
   io.sockets.emit('state', players);
-
   if (map.locationReserveChanged) {
     map.locationReserveChanged = false;
     io.sockets.emit('locationReservations', map.locationReservations);
   }
-}, 1000 / 60);
+  setTimeout(run, 1000 / 60);
+}
 
-setInterval(() => {
-  map.saveData(saveDir);
-}, 60 * 1000);
+setInterval(function () {
+  if (!map.data) {
+    createData(map);
+  }
+  updateData(map);
+  io.sockets.emit('data', map.data);
+}, 1000);
 
+// setInterval(() => {
+//   map.saveData(saveDir);
+//   // console.log(JSON.stringify(map.data));
+// }, 60 * 1000);
+
+run();
 module.exports = io;
