@@ -1,5 +1,4 @@
 function AStarPath(position, direction, distance) {
-    // this.id = `city:${cityIndex}x:${x}y:${y}`;
     this.position = position;
     this.direction = direction;
     this.distance = distance;
@@ -7,10 +6,10 @@ function AStarPath(position, direction, distance) {
 
 AStarPath.prototype.moveAgent = function (agent, deltaTime) {
     let moveAmount = agent.speed * deltaTime;
-    console.log(`current position: ${printPos(agent.position)}`);
+    // console.log(`current position: ${printPos(agent.position)}`);
     agent.position.x += this.direction.x * moveAmount;
     agent.position.y += this.direction.y * moveAmount;
-    console.log(`new position: ${printPos(agent.position)}`);
+    // console.log(`new position: ${printPos(agent.position)}`);
     
     //clamp
     if (this.direction.x < 0) {
@@ -23,7 +22,7 @@ AStarPath.prototype.moveAgent = function (agent, deltaTime) {
     } else if (this.direction.y > 0) {
         agent.position.y = Math.min(this.position.y, agent.position.y);
     }
-    console.log(`clamped position: ${printPos(agent.position)}`);
+    // console.log(`clamped position: ${printPos(agent.position)}`);
 
     //return distance to target
     return distance(this.position,agent.position);
